// backend/api/migrations/snapshots/ed10452ecc0bf3f47a88b45953058ed2a2cbe204ba1e7556cf936104ac2e1ea0/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'USER' | 'CREATOR' | 'ARTIST' | 'MANAGER' | 'LABEL' | 'ADMIN' | 'OWNER';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'USER' | 'CREATOR' | 'ARTIST' | 'MANAGER' | 'LABEL' | 'ADMIN' | 'OWNER';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: 'USER' | 'CREATOR' | 'ARTIST' | 'MANAGER' | 'LABEL' | 'ADMIN' | 'OWNER';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: 'USER' | 'CREATOR' | 'ARTIST' | 'MANAGER' | 'LABEL' | 'ADMIN' | 'OWNER';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;